refactor: check layout folder 이동
